pyproject.toml
[tool.poetry]
name = "robin"
version = "0.1.0"
description = "ROBIN proposes an AI-driven, human-centric robotic system that leverages ROS 2 and FIWARE to enable adaptive processes, data standardization, and enhanced operator interaction."
authors = ["dhaasb <68024961+Daniel-Haas-B@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pydantic = "^2.11.7"
requests = "^2.32.5"
typer = "^0.17.3"
numpy = "^2.3.2"
matplotlib = "^3.10.6"
pep257 = "^0.7.0"
torch = {version = "^2.5.1", source = "pytorch-cpu"}
pandas = "^2.3.3"
httpx = "^0.28.1"
pyyaml = "^6.0.3"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.1.0"
sphinx-rtd-theme = "^1.3.0"
furo = "^2024.8.6"
myst-parser = "^2.0.0"
sphinx-autodoc-typehints = "^1.24.0"
sphinx-copybutton = "^0.5.2"
sphinxcontrib-mermaid = "^0.9.2"
sphinx-tabs = "^3.4.1"
sphinx-design = "^0.5.0"
linkify-it-py = "^2.0.3"

[tool.poetry.group.dev.dependencies]
# Add development dependencies here when needed
pre-commit = "^4.2.0"
flake8 = "^7.3.0"
pytest = "^8.4.1"
httpx = "^0.28.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^7.0.0"
selenium = "^4.41.0"
webdriver-manager = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
